+++ packages/data/src/types/abstract-query/fields/field-nodes.ts
@@ -1,5 +1,5 @@
-import type { AbstractQueryFieldNodePrimitive } from './primitive.js';
 import type { AbstractQueryFieldNodeFn } from './function.js';
+import type { AbstractQueryFieldNodePrimitive } from './primitive.js';
 import type { AbstractQueryFieldNodeRelated } from './related.js';
 
 export type AbstractQueryFieldNode =
