+++ packages/data/src/types/abstract-query/fields/index.ts
@@ -1,0 +1,4 @@
+export * from './field-nodes.js';
+export * from './function.js';
+export * from './primitive.js';
+export * from './related.js';
