+++ packages/data/src/types/abstract-query/fields/related.ts
@@ -1,9 +1,9 @@
 /**
  * This file will be split up into multiple files soon.
  */
-import type { AbstractQueryModifiers } from '../modifiers/index.js';
-import type { AbstractQueryFieldNodePrimitive } from './primitive.js';
+import type { AbstractQueryModifiers } from '../modifiers.js';
 import type { AbstractQueryFieldNodeFn } from './function.js';
+import type { AbstractQueryFieldNodePrimitive } from './primitive.js';
 
 /**
  * This is a basic interface for all relational field types.
