+++ packages/data/src/types/abstract-query/modifiers/filters.ts
@@ -1,0 +1,5 @@
+import type { AbstractQueryConditionNode } from './filters/conditions.js';
+import type { AbstractQueryNodeLogical } from './filters/logical.js';
+import type { AbstractQueryNodeNegate } from './filters/negate.js';
+
+export type AbstractQueryFilterNode = AbstractQueryConditionNode | AbstractQueryNodeLogical | AbstractQueryNodeNegate;
