+++ packages/data/src/types/abstract-query/modifiers/filters/conditions/geo-condition-bbox.ts
@@ -1,4 +1,5 @@
 import type { GeoJSONGeometryCollection, GeoJSONMultiPolygon, GeoJSONPolygon } from 'wellknown';
+
 import type { AbstractQueryFieldNodePrimitive } from '../../../fields/primitive.js';
 
 /**
