+++ packages/data/src/types/abstract-query/modifiers/filters/index.ts
@@ -1,9 +1,4 @@
-import type { AbstractQueryConditionNode } from './conditions/index.js';
-import type { AbstractQueryNodeLogical } from './logical.js';
-import type { AbstractQueryNodeNegate } from './negate.js';
-
+export * from './conditions.js';
 export * from './conditions/index.js';
 export * from './logical.js';
 export * from './negate.js';
-
-export type AbstractQueryFilterNode = AbstractQueryConditionNode | AbstractQueryNodeLogical | AbstractQueryNodeNegate;
