+++ packages/data/src/types/index.ts
@@ -1,2 +1,3 @@
-export * from './abstract-query/abstract-query.js';
+export * from './abstract-query.js';
+export * from './abstract-query/index.js';
 export * from './driver.js';
